+++ frontend/src/metabase/common/components/EntityPicker/SpecificEntityPickers/CollectionPicker.tsx
@@ -5,8 +5,7 @@ import { CollectionsApi, UserApi } from "metabase/services";
 
 import Search from "metabase/entities/search";
 
-import { LoadingSpinner } from "../components/LoadingSpinner";
-import { NestedItemPicker } from "../NestedItemPicker";
+import { LoadingSpinner, NestedItemPicker } from "../components";
 import type { PickerState } from "../types";
 
 type SearchCollection = Partial<Collection> & { model: string };
@@ -18,18 +17,12 @@ interface CollectionPickerProps {
 }
 
 function getCollectionIdPath(collection: Collection) {
-  const pathFromRoot = collection?.location?.split("/").filter(Boolean).map(Number) ?? [];
+  const pathFromRoot =
+    collection?.location?.split("/").filter(Boolean).map(Number) ?? [];
 
-  const path = collection.is_personal ? [
-    null,
-    ...pathFromRoot,
-    collection.id,
-  ] : [
-    null,
-    'root',
-    ...pathFromRoot,
-    collection.id,
-  ];
+  const path = collection.is_personal
+    ? [null, ...pathFromRoot, collection.id]
+    : [null, "root", ...pathFromRoot, collection.id];
 
   return path;
 }
@@ -47,69 +40,79 @@ export function CollectionPicker({
   value,
   options = defaultOptions,
 }: CollectionPickerProps) {
-  const [ initialState, setInitialState ] = useState<PickerState<Collection>>();
+  const [initialState, setInitialState] = useState<PickerState<Collection>>();
 
-  const onFolderSelect = useCallback(async (folder?: SearchCollection): Promise<SearchCollection[]> => {
-    if (!folder?.id) {
-      const ourAnalytics = await CollectionsApi.getRoot();
+  const onFolderSelect = useCallback(
+    async (folder?: SearchCollection): Promise<SearchCollection[]> => {
+      if (!folder?.id) {
+        const ourAnalytics = await CollectionsApi.getRoot();
 
-      const collectionsData = [{
-        ...ourAnalytics,
-        model: 'collection',
-        id: 'root',
-      }];
+        const collectionsData = [
+          {
+            ...ourAnalytics,
+            model: "collection",
+            id: "root",
+          },
+        ];
 
-      if (options.showPersonalCollection) {
-        const currentUser = await UserApi.current();
-        const personalCollection = await CollectionsApi.get({
-           id: currentUser.personal_collection_id
-        });
-        collectionsData.push({
-          ...personalCollection,
-          model: 'collection',
-        });
+        if (options.showPersonalCollection) {
+          const currentUser = await UserApi.current();
+          const personalCollection = await CollectionsApi.get({
+            id: currentUser.personal_collection_id,
+          });
+          collectionsData.push({
+            ...personalCollection,
+            model: "collection",
+          });
+        }
+
+        return collectionsData;
       }
 
-      return collectionsData;
-    }
+      // because folders are also selectable items in the collection picker, we always select the folder
+      onItemSelect(folder ?? { id: "root", model: "collection" });
 
-    // because folders are also selectable items in the collection picker, we always select the folder
-    onItemSelect(folder ?? { id: 'root', model: 'collection' });
+      const items = await Search.api.list({
+        collection: folder.id,
+        models: ["collection"],
+      });
 
-    const items = await Search.api.list({
-      collection: folder.id,
-      models: ['collection'],
-    });
-
-    return items.data;
-  }, [onItemSelect, options]);
+      return items.data;
+    },
+    [onItemSelect, options],
+  );
 
   useEffect(() => {
     if (value?.id) {
-      CollectionsApi
-        .get({ id: value.id })
-        .then(async (collection) => {
-          const path = getCollectionIdPath(collection);
+      CollectionsApi.get({ id: value.id }).then(async collection => {
+        const path = getCollectionIdPath(collection);
 
-          const stack = await Promise.all(
-            path.map(async (id, index) => {
-              const items = await onFolderSelect({ id: (id as CollectionId), model: 'collection' });
-              const selectedItem = items.find(item => item.id === path[index + 1]) ?? null;
-              return({
-                items,
-                selectedItem,
-              });
-            }),
-          );
+        const stack = await Promise.all(
+          path.map(async (id, index) => {
+            const items = await onFolderSelect({
+              id: id as CollectionId,
+              model: "collection",
+            });
+            const selectedItem =
+              items.find(item => item.id === path[index + 1]) ?? null;
+            return {
+              items,
+              selectedItem,
+            };
+          }),
+        );
 
-          setInitialState(stack);
-        });
+        setInitialState(stack);
+      });
     } else {
       // default to showing our analytics selected
-      onFolderSelect().then( async (items) => {
+      onFolderSelect().then(async items => {
         setInitialState([
-          { items, selectedItem: { id: 'root', model: 'collection' } },
-          { items: await onFolderSelect({ id: 'root', model: 'collection' }), selectedItem: null },
+          { items, selectedItem: { id: "root", model: "collection" } },
+          {
+            items: await onFolderSelect({ id: "root", model: "collection" }),
+            selectedItem: null,
+          },
         ]);
       });
     }
