+++ frontend/src/metabase/common/components/EntityPicker/SpecificEntityPickers/QuestionPicker.tsx
@@ -4,8 +4,7 @@ import { GET } from "metabase/lib/api";
 import type { Collection } from "metabase-types/api";
 import type Question from "metabase-lib/Question";
 
-import { NestedItemPicker } from "../NestedItemPicker";
-
+import { NestedItemPicker } from "../components";
 
 interface QuestionPickerProps {
   onItemSelect: (item: Question) => void;
@@ -17,16 +16,25 @@ const collectionList = GET("/api/collection/:collection/items");
 const collectionAPI = GET("/api/collection/:id");
 
 const sortFoldersFirst = (a: any, b: any) => {
-  return a.model === 'collection' ? -1 : 1;
-}
+  return a.model === "collection" ? -1 : 1;
+};
 
-export function QuestionPicker({ onItemSelect, initialCollectionId }: QuestionPickerProps) {
-  const [ initialState, setInitialState ] = useState<any>();
+export function QuestionPicker({
+  onItemSelect,
+  initialCollectionId,
+}: QuestionPickerProps) {
+  const [initialState, setInitialState] = useState<any>();
 
   const onFolderSelect = async (folder: Collection) => {
     const items = !folder
-     ? await collectionList({collection: 'root' }, {model: ["question", "collection" ]})
-     : await collectionList({collection: folder.id }, { model: ["question", "collection" ]});
+      ? await collectionList(
+          { collection: "root" },
+          { model: ["question", "collection"] },
+        )
+      : await collectionList(
+          { collection: folder.id },
+          { model: ["question", "collection"] },
+        );
 
     return items.data.sort(sortFoldersFirst) as Collection[];
   };
@@ -35,19 +43,25 @@ export function QuestionPicker({ onItemSelect, initialCollectionId }: QuestionPi
     if (initialCollectionId) {
       // FIXME, if the initialCollectionID changes, we'll do all these net requests,
       // but it won't change anything in entity picker
-      collectionAPI({ id: initialCollectionId }).then(async (collection) => {
-        const path = ['root', ...collection.location.split('/'), initialCollectionId].filter(Boolean);
+      collectionAPI({ id: initialCollectionId }).then(async collection => {
+        const path = [
+          "root",
+          ...collection.location.split("/"),
+          initialCollectionId,
+        ].filter(Boolean);
 
-        const stack = await Promise.all(path.map(async (id, index) => {
-          return {
-            items: (await onFolderSelect({ id })).sort(sortFoldersFirst),
-            selectedId: path[index+1] ?? null
-          }
-        }));
+        const stack = await Promise.all(
+          path.map(async (id, index) => {
+            return {
+              items: (await onFolderSelect({ id })).sort(sortFoldersFirst),
+              selectedId: path[index + 1] ?? null,
+            };
+          }),
+        );
         setInitialState(stack);
       });
     } else {
-      onFolderSelect({ id: 'root' }).then((items) => {
+      onFolderSelect({ id: "root" }).then(items => {
         items.sort(sortFoldersFirst);
         setInitialState([{ items, selectedId: null }]);
       });
@@ -66,5 +80,5 @@ export function QuestionPicker({ onItemSelect, initialCollectionId }: QuestionPi
       onItemSelect={onItemSelect}
       initialState={initialState}
     />
-  )
+  );
 }
