+++ frontend/src/metabase/common/components/EntityPicker/components/ButtonBar/index.ts
@@ -1,0 +1,1 @@
+export * from "./ButtonBar";
