+++ frontend/src/metabase/common/components/EntityPicker/components/ItemList/ItemList.styled.tsx
@@ -1,0 +1,7 @@
+import styled from "@emotion/styled";
+import { Box } from "metabase/ui";
+
+export const PickerColumn = styled(Box)`
+  flex-basis: 310px;
+  padding: 1rem 1rem 1rem 1.5rem;
+`;
