+++ frontend/src/metabase/common/components/EntityPicker/components/ItemList/index.ts
@@ -1,0 +1,1 @@
+export * from "./ItemList";
