+++ frontend/src/metabase/common/components/EntityPicker/components/LoadingSpinner/index.ts
@@ -1,0 +1,1 @@
+export * from "./LoadingSpinner"
