+++ frontend/src/metabase/common/components/EntityPicker/components/NestedItemPicker/index.ts
@@ -1,0 +1,1 @@
+export * from "./NestedItemPicker"
