+++ frontend/src/metabase/common/components/EntityPicker/components/SinglePickerView/index.ts
@@ -1,0 +1,1 @@
+export * from "./SinglePickerView";
