+++ frontend/src/metabase/common/components/EntityPicker/components/TabsView/index.ts
@@ -1,0 +1,1 @@
+export * from "./TabsView";
