+++ frontend/src/metabase/common/components/EntityPicker/components/index.ts
@@ -1,0 +1,5 @@
+export * from "./ButtonBar";
+export * from "./LoadingSpinner";
+export * from "./NestedItemPicker";
+export * from "./TabsView";
+export * from "./SinglePickerView";
