+++ frontend/src/metabase/common/components/EntityPicker/containers/EntityPickerModal/index.ts
@@ -1,0 +1,1 @@
+export * from "./EntityPickerModal";
