+++ frontend/src/metabase/common/components/EntityPicker/containers/index.ts
@@ -1,0 +1,1 @@
+export * from "./EntityPickerModal";
