+++ frontend/src/metabase/common/components/EntityPicker/index.ts
@@ -1,2 +1,2 @@
-export * from "./NestedItemPicker";
-export * from "./EntityPickerModal";
+export * from "./components/NestedItemPicker/NestedItemPicker";
+export * from "./containers/EntityPickerModal/EntityPickerModal";
