+++ frontend/src/metabase/common/components/EntityPicker/types.ts
@@ -2,5 +2,12 @@ export type PickerState<T> = PickerStateItem<T>[];
 
 export type PickerStateItem<T> = {
   items: (Partial<T> & { model: string })[];
-  selectedItem: Partial<T> & { model: string } | null;
+  selectedItem: (Partial<T> & { model: string }) | null;
+};
+
+export type EntityPickerModalOptions = {
+  showPersonalCollection?: boolean;
+  showSearch?: boolean;
+  showRecents?: boolean;
+  hasConfirmButtons?: boolean;
 };
