+++ frontend/src/metabase/home/components/HomePage/HomePage.tsx
@@ -16,7 +16,7 @@ import { canUseMetabotOnDatabase } from "metabase/metabot/utils";
 import type { CollectionItem, DashboardId } from "metabase-types/api";
 import { getSettingsLoading } from "metabase/selectors/settings";
 
-// import { EntityPickerModal } from "metabase/common/components/EntityPicker";
+import { EntityPickerModal } from "metabase/common/components/EntityPicker";
 import type Database from "metabase-lib/metadata/Database";
 import {
   getCustomHomePageDashboardId,
@@ -40,7 +40,7 @@ export const HomePage = (): JSX.Element => {
 
   return (
     <HomeLayout hasMetabot={getHasMetabot(databases, models, isMetabotEnabled)}>
-      {/* <EntityPickerModal onItemSelect={console.log} tabs={["collection"]} /> */}
+      <EntityPickerModal onItemSelect={console.log} tabs={["collection"]} />
       <HomeContent />
     </HomeLayout>
   );
